Split zebra_user.py
Move all functions in zebra_user.py that can be reused in other load
testing files to onadata_libs.py.
