+++ requirements.txt
@@ -14,6 +14,7 @@ gevent==1.2.2             # via locustio
 greenlet==0.4.12          # via gevent
 idna==2.6                 # via requests
 isort==4.2.15             # via pylint
+yapf==0.17.0
 itsdangerous==0.24        # via flask
 jinja2==2.9.6             # via flask
 lazy-object-proxy==1.3.1  # via astroid
